[CHERRY-PICK] Move edit_entry_with_cmd() to libimagutil
