+++ libimagentryedit/Cargo.toml
@@ -4,7 +4,6 @@ version = "0.2.0"
 authors = ["Matthias Beyer <mail@beyermatthias.de>"]
 
 [dependencies]
-tempfile = "2.1.1"
 
 [dependencies.libimagerror]
 path = "../libimagerror"
@@ -14,4 +13,7 @@ path = "../libimagrt"
 
 [dependencies.libimagstore]
 path = "../libimagstore"
+
+[dependencies.libimagutil]
+path = "../libimagutil"
