+++ libimagentryedit/src/edit.rs
@@ -39,44 +39,20 @@ impl<'a> Edit for FileLockEntry<'a> {
 }
 
 pub fn edit_in_tmpfile(rt: &Runtime, s: &mut String) -> Result<()> {
-    edit_in_tmpfile_with_command(rt.editor(), s)
-}
+    use libimagutil::edit::edit_in_tmpfile_with_command;
 
-pub fn edit_in_tmpfile_with_command(cmd: Command, s: &mut String) -> Result<()> {
-    use tempfile::NamedTempFile;
-    use std::io::Seek;
-    use std::io::Read;
-    use std::io::SeekFrom;
-    use std::io::Write;
-
-    let file      = try!(NamedTempFile::new().map_err_into(EditErrorKind::IOError));
-    let file_path = file.path();
-    let mut file  = try!(file.reopen().map_err_into(EditErrorKind::IOError));
-
-    try!(file.write_all(&s.clone().into_bytes()[..]).map_err_into(EditErrorKind::IOError));
-    try!(file.sync_data().map_err_into(EditErrorKind::IOError));
-
-    if let Some(mut editor) = rt.editor() {
-        let exit_status = editor.arg(file_path).status();
-
-        match exit_status.map(|s| s.success()).map_err(Box::new) {
-            Ok(true)  => {
-                file.sync_data()
-                    .and_then(|_| file.seek(SeekFrom::Start(0)))
-                    .and_then(|_| {
-                        let mut new_s = String::new();
-                        let res = file.read_to_string(&mut new_s);
-                        *s = new_s;
-                        res
-                    })
-                    .map(|_| ())
-                    .map_err_into(EditErrorKind::IOError)
-            },
-            Ok(false) => Err(EditErrorKind::ProcessExitFailure.into()),
-            Err(e)    => Err(EditErrorKind::IOError.into_error_with_cause(e)),
-        }
-    } else {
-        Err(EditErrorKind::InstantiateError.into())
-    }
+    rt.editor()
+        .ok_or(EditErrorKind::NoEditor.into_error())
+        .and_then(|editor| {
+            edit_in_tmpfile_with_command(editor, s)
+                .map_err_into(EditErrorKind::IOError)
+                .and_then(|worked| {
+                    if !worked {
+                        Err(EditErrorKind::ProcessExitFailure.into())
+                    } else {
+                        Ok(())
+                    }
+                })
+        })
 }
