+++ libimagentryedit/src/error.rs
@@ -1,6 +1,7 @@
 generate_error_module!(
     generate_error_types!(EditError, EditErrorKind,
         IOError             => "IO Error",
+        NoEditor            => "No editor set",
         ProcessExitFailure  => "Process did not exit properly",
         InstantiateError    => "Instantation error"
     );
