+++ libimagentryedit/src/lib.rs
@@ -1,7 +1,7 @@
 #[macro_use] extern crate libimagerror;
 extern crate libimagstore;
 extern crate libimagrt;
-extern crate tempfile;
+extern crate libimagutil;
 
 pub mod edit;
 pub mod error;
